# Improve handling of uncorrelated subselects