--- a/src/lib/operators/table_scan.cpp
+++ b/src/lib/operators/table_scan.cpp
@@ -234,6 +234,18 @@ std::unique_ptr<AbstractTableScanImpl> TableScan::create_impl() const {
       return std::make_unique<ColumnVsColumnTableScanImpl>(input_table_left(), left_column_expression->column_id,
                                                            predicate_condition, right_column_expression->column_id);
     }
+
+    // Predicate pattern: <column> <binary predicate_condition> <uncorrelated subquery> // TODO Reversed
+    if (const auto right_subquery =
+            std::dynamic_pointer_cast<PQPSelectExpression>(right_operand); left_column_expression && !right_subquery->is_correlated()) {
+      auto uncorrelated_select_results = ExpressionEvaluator::populate_uncorrelated_select_results_cache({right_subquery});
+      auto table = uncorrelated_select_results->begin()->second;
+      Assert(table->row_count() == 1, "Expected subquery to return a single row");  // TODO AssertInput?
+      Assert(table->column_count() == 1, "Expected subquery to return a single row");  // TODO AssertInput?
+      auto result = table->get_chunk(ChunkID{0})->get_segment(ColumnID{0})->operator[](0);
+      return std::make_unique<ColumnVsValueTableScanImpl>(input_table_left(), left_column_expression->column_id,
+                                                          predicate_condition, result);
+    }
   }
 
   if (const auto is_null_expression = std::dynamic_pointer_cast<IsNullExpression>(_predicate)) {