--- a/src/lib/operators/table_scan/column_like_table_scan_impl.cpp
+++ b/src/lib/operators/table_scan/column_like_table_scan_impl.cpp
@@ -27,7 +27,7 @@ ColumnLikeTableScanImpl::ColumnLikeTableScanImpl(const std::shared_ptr<const Tab
       _matcher{pattern},
       _invert_results(predicate_condition == PredicateCondition::NotLike) {}
 
-std::string ColumnLikeTableScanImpl::description() const { return "LikeScan"; }
+std::string ColumnLikeTableScanImpl::description() const { return "ColumnLike"; }
 
 void ColumnLikeTableScanImpl::_scan_non_reference_segment(const BaseSegment& segment, const ChunkID chunk_id,
                                                           PosList& matches,