--- a/src/lib/operators/table_scan/column_vs_column_table_scan_impl.cpp
+++ b/src/lib/operators/table_scan/column_vs_column_table_scan_impl.cpp
@@ -24,7 +24,7 @@ ColumnVsColumnTableScanImpl::ColumnVsColumnTableScanImpl(const std::shared_ptr<c
       _predicate_condition(predicate_condition),
       _right_column_id{right_column_id} {}
 
-std::string ColumnVsColumnTableScanImpl::description() const { return "ColumnComparison"; }
+std::string ColumnVsColumnTableScanImpl::description() const { return "ColumnVsColumn"; }
 
 std::shared_ptr<PosList> ColumnVsColumnTableScanImpl::scan_chunk(ChunkID chunk_id) const {
   const auto chunk = _in_table->get_chunk(chunk_id);