--- a/src/test/expression/expression_test.cpp
+++ b/src/test/expression/expression_test.cpp
@@ -87,6 +87,13 @@ TEST_F(ExpressionTest, DeepEquals) {
   EXPECT_EQ(*case_c, *case_c);
   EXPECT_NE(*case_a, *case_b);
   EXPECT_NE(*case_a, *case_c);
+
+  const auto parameter_a = correlated_parameter_(ParameterID{5}, a);
+  const auto parameter_b = correlated_parameter_(ParameterID{5}, a);
+  EXPECT_EQ(*parameter_a, *parameter_b);
+  static_cast<CorrelatedParameterExpression&>(*parameter_a).set_value(3);
+  static_cast<CorrelatedParameterExpression&>(*parameter_b).set_value(4);
+  EXPECT_NE(*parameter_a, *parameter_b);
 }
 
 TEST_F(ExpressionTest, DeepCopy) {
@@ -95,6 +102,13 @@ TEST_F(ExpressionTest, DeepCopy) {
 
   const auto expr_b = and_(greater_than_equals_(15, 12), or_(greater_than_(5, 3), less_than_(3, 5)));
   EXPECT_EQ(*expr_b, *expr_b->deep_copy());
+
+  const auto parameter_a = correlated_parameter_(ParameterID{5}, a);
+  static_cast<CorrelatedParameterExpression&>(*parameter_a).set_value(3);
+  const auto parameter_b = parameter_a->deep_copy();
+  EXPECT_EQ(*parameter_a, *parameter_b);
+  static_cast<CorrelatedParameterExpression&>(*parameter_b).set_value(4);
+  EXPECT_NE(*parameter_a, *parameter_b);
 }
 
 TEST_F(ExpressionTest, RequiresCalculation) {