For queries like `... WHERE x = (SELECT ...)`, we currently use the ExpressionEvaluator. If, however, the subselect is non-correlated, it might make more sense to evaluate the subselect first and then use a regular scan (which is better optimized).

Unfortunately, I did not see any benefits in the TPC-H. While it matches queries 11, 15, and 22, the input sizes for the table scan are too small for it to make a difference. But now that it's already implemented, here is a synthetic query to make up for it:

```
SELECT COUNT(*) FROM lineitem WHERE l_partkey = (SELECT MAX(p_partkey) FROM part)

Before: 151 ms 843 µs
Now: 75 ms 55 µs
```