@mrks 2019-01-11
How about this? I like this approach better, and it also handles BETWEEN etc.