@mrks 2019-01-15
Should be all addressed. I had to fix what I consider a bug in the correlated predicate expression, please have a look.